Finished part 1 classification code
Updated the clang format column limit to be less restrictive, so a few files have been reformatted without changes to the code

Added figures to the report target, so now part 1 figures can be generated using the report target

Fixed a potential future bug with the eigen library by requiring all STL vectors of Eigen vectors/matrices use a special allocator that respects alignment requirements

Fixed a bug where the covariance matrix for sample 2 from data set B was incorrect (used std. dev. instead of variance)

Added the option to output misclassified points to a plot file. Use the '-pmN' switches where N is 1-2.

Fixed a bug where the discriminant functions were passed the covariance matrix instead of the inverse of the covariance matrix.

Added the option to output the parameters to the decision boundary to a file. Parameters are named a letter of the alphabet starting with a and increasing based on 'alphabetic' order of the terms they are attached to. Since the boundary is a quadric, this looks something like ax^2 + bxy + cy^2 + dx + ey + f. There are also parameters for the window size on which to plot this boundary if you are plotting it on the same plot as the observations.

Changed the plots so that the misclassification plots and the whole sample plots are always plotted on the same window, so they can overlap. Also removed the transparent background for easier viewing in places with a non-white background.